removing globals
