+++ src/CMakeLists.txt
@@ -7,3 +7,4 @@ add_subdirectory(cmd-insert)
 add_subdirectory(cmd-c)
 add_subdirectory(cmd-ld)
 add_subdirectory(libannotatesyscalls)
+add_subdirectory(libmodifysource)
