+++ src/cmd-insert/CMakeLists.txt
@@ -8,4 +8,4 @@ llvm_map_components_to_libnames(llvm_libs support core)
 
 find_package(clang REQUIRED)
 
-target_link_libraries(cmd-insert libannotatesyscalls ${llvm_libs} clangTooling)
+target_link_libraries(cmd-insert libannotatesyscalls libmodifysource ${llvm_libs} clangTooling)
